requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-braces==1.15.0
django-model-utils==4.2.0
django-phone-field==1.8.1
django-star-ratings==0.9.2
django-tinymce==3.4.0
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
swapper==1.3.0